+++ g2engineconfigurationjson/g2engineconfigurationjson_linux.go
@@ -6,33 +6,29 @@ package g2engineconfigurationjson
 // Internal methods
 // ----------------------------------------------------------------------------
 
-func buildStruct(specificDatabaseUrl string, licenseStringBase64 string) G2Configuration {
-
+// func buildStruct(specificDatabaseUrl string, licenseStringBase64 string, senzingDirectory string) G2Configuration {
+func buildStruct(attributeMap map[string]string) G2Configuration {
 	var result G2Configuration
 
-	if len(licenseStringBase64) > 0 {
-		result = G2Configuration{
-			Pipeline: G2ConfigurationPipeline{
-				ConfigPath:          "/etc/opt/senzing",
-				LicenseStringBase64: licenseStringBase64,
-				ResourcePath:        "/opt/senzing/g2/resources",
-				SupportPath:         "/opt/senzing/data",
-			},
-			Sql: G2ConfigurationSql{
-				Connection: specificDatabaseUrl,
-			},
-		}
-	} else {
-		result = G2Configuration{
-			Pipeline: G2ConfigurationPipeline{
-				ConfigPath:   "/etc/opt/senzing",
-				ResourcePath: "/opt/senzing/g2/resources",
-				SupportPath:  "/opt/senzing/data",
-			},
-			Sql: G2ConfigurationSql{
-				Connection: specificDatabaseUrl,
-			},
-		}
+	databaseUrl, ok := attributeMap["databaseUrl"]
+	if !ok {
+		return result
+	}
+
+	result = G2Configuration{
+		Pipeline: G2ConfigurationPipeline{
+			ConfigPath:   "/etc/opt/senzing",
+			ResourcePath: "/opt/senzing/g2/resources",
+			SupportPath:  "/opt/senzing/data",
+		},
+		Sql: G2ConfigurationSql{
+			Connection: databaseUrl,
+		},
+	}
+
+	licenseStringBase64, ok := attributeMap["licenseStringBase64"]
+	if ok {
+		result.Pipeline.LicenseStringBase64 = licenseStringBase64
 	}
 
 	return result